Stopped the NPC at couch for 5 seconds (Issue #87)
Added a timer to count for how long we want the NPC to pause at the couch for.

Also added a couch and stool "reached" boolean which becomes true if the NPC has reached the point. This is necessary because unity changes the value of the distance between the NPC and the point once the NPC stops walking. So in one frame it can pass the if condition saying it has reached the point, but in another frame, it can fail, even though the NPC hasn't moved.