Player can give bike to npc next to taxi (Issue #150)
Fixed the bug where if you gave the NPC the bike as he reaches the taxi, he'll appear with a bike the next time he exits the building.

Now, if you give the npc the bike, he will ride it off into the distance and the task will be successful. Adds a 2 second margin where you can still give the npc the bike.