Backyard bin NPC now launches every time (Issue #115)
Reset all variables related to the NPCs position and status (filled bin and completed bools) so that it was ready to be launched as soon as it was near the point.

This was working before but not 100% of the time. Very rarely, the NPC would not launch with the task. This almost never happened, but now it can't happen.